+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+- Add `remember openai profile` command to remember the pretext of AI prompt.
+- Add `show openai profile` command to show the pretext of AI prompt.
 - Make `OPENAI_CHAT_PRETEXT` optional.
 
 ## [0.1.0] - 2022-12-17